begin of refactor
